fixed the problem with correct calculating of amount

+++ src/main/java/com/project/carsharingapp/service/PaymentAmountHandler.java
@@ -4,7 +4,7 @@ import com.project.carsharingapp.model.Payment;
 import java.math.BigDecimal;
 
 public interface PaymentAmountHandler {
-    Long getPaymentAmount(BigDecimal dailyFee, int numberOfDays);
+    BigDecimal getPaymentAmount(BigDecimal dailyFee, int numberOfDays);
 
     boolean isApplicable(Payment.Type type);
 }
